refactor: merge DrawerOneUIExtensions and DeleteAppDataUtils into their natural homes
- Fold DrawerOneUIExtensions.kt into DrawerUtils.kt (same concern, no split needed)
- Move deleteAppDataAndExit from DeleteAppDataUtils.kt into PreferenceUtils.kt
  (only ever called from there)
- Update Kover exclusions to reference the new class/method locations
